+++ config/GatewayIntentBits.js
@@ -8,6 +8,8 @@ const { GatewayIntentBits } = require("discord.js");
 const intents = [
   GatewayIntentBits.Guilds,
   GatewayIntentBits.GuildMessages,
+  GatewayIntentBits.GuildMessageReactions,
+  GatewayIntentBits.MessageContent,
 ];
 
 module.exports = {